requirements.txt
fastapi
uvicorn
openai
sentence-transformers
faiss-cpu
python-dotenv
pymupdf
pandas
python-multipart
typing
neo4j